Open the widgets in a bottom panel instead of floating centered
The sidebar version works a lot different though. It needs to be toggled
on. But it creates another version if I run it again. The right way to
do this might be to have it just created at startup and then toggled on
or off with the keybind. I'll have to mess with it some depending what I
end up actually using.